feat(endpoints): centralised all endpoints with types
