+++ packages/core/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/lit-core';
+export * from './lib/endpoint-version';
