+++ packages/lit-node-client-nodejs/src/lib/lit-node-client-nodejs.ts
@@ -19,6 +19,7 @@ import {
   AuthMethodType,
   EITHER_TYPE,
   LIT_ACTION_IPFS_HASH,
+  LIT_ENDPOINT,
   LIT_ERROR,
   LIT_SESSION_KEY_URI,
   LOCAL_STORAGE_KEYS,
@@ -26,7 +27,7 @@ import {
   SIGTYPE,
   SIWE_DELEGATION_URI,
 } from '@lit-protocol/constants';
-import { LitCore } from '@lit-protocol/core';
+import { LitCore, composeLitUrl } from '@lit-protocol/core';
 import {
   checkSevSnpAttestation,
   combineEcdsaShares,
@@ -732,8 +733,10 @@ export class LitNodeClientNodeJs
     logWithRequestId(requestId, 'getJsExecutionShares');
 
     // -- execute
-
-    let urlWithPath = `${url}/web/execute`;
+    const urlWithPath = composeLitUrl({
+      url,
+      endpoint: LIT_ENDPOINT.EXECUTE_JS,
+    });
 
     if (!authSig) {
       throw new Error('authSig or sessionSig is required');
@@ -765,7 +768,10 @@ export class LitNodeClientNodeJs
     requestId: string
   ) => {
     logWithRequestId(requestId, 'getPkpSigningShares');
-    const urlWithPath = `${url}/web/pkp/sign`;
+    const urlWithPath = composeLitUrl({
+      url,
+      endpoint: LIT_ENDPOINT.PKP_SIGN,
+    });
     if (!params.authSig) {
       throw new Error('authSig is required');
     }
@@ -783,7 +789,10 @@ export class LitNodeClientNodeJs
     requestId: string
   ) => {
     logWithRequestId(requestId, 'getPkpSigningShares');
-    const urlWithPath = `${url}/web/pkp/claim`;
+    const urlWithPath = composeLitUrl({
+      url,
+      endpoint: LIT_ENDPOINT.PKP_CLAIM,
+    });
     if (!params.authMethod) {
       throw new Error('authMethod is required');
     }
@@ -810,7 +819,11 @@ export class LitNodeClientNodeJs
     requestId: string
   ): Promise<NodeCommandResponse> => {
     logWithRequestId(requestId, 'getSigningShareForToken');
-    const urlWithPath = `${url}/web/signing/access_control_condition`;
+
+    const urlWithPath = composeLitUrl({
+      url,
+      endpoint: LIT_ENDPOINT.SIGN_ACCS,
+    });
 
     return this.sendCommandToNode({
       url: urlWithPath,
@@ -834,7 +847,11 @@ export class LitNodeClientNodeJs
     requestId: string
   ): Promise<NodeCommandResponse> => {
     log('getSigningShareForDecryption');
-    const urlWithPath = `${url}/web/encryption/sign`;
+
+    const urlWithPath = composeLitUrl({
+      url,
+      endpoint: LIT_ENDPOINT.ENCRYPTION_SIGN,
+    });
 
     return await this.sendCommandToNode({
       url: urlWithPath,
@@ -862,7 +879,11 @@ export class LitNodeClientNodeJs
       id: string
     ): Promise<SuccessNodePromises<any> | RejectedNodePromises> => {
       log('signConditionEcdsa');
-      const urlWithPath = `${url}/web/signing/signConditionEcdsa`;
+
+      const urlWithPath = composeLitUrl({
+        url,
+        endpoint: LIT_ENDPOINT.SIGN_ECDSA,
+      });
 
       const data = {
         access_control_conditions: params.accessControlConditions,
@@ -2693,7 +2714,10 @@ export class LitNodeClientNodeJs
     requestId: string
   ) => {
     log('getSignSessionKeyShares');
-    const urlWithPath = `${url}/web/sign_session_key`;
+    const urlWithPath = composeLitUrl({
+      url,
+      endpoint: LIT_ENDPOINT.SIGN_SESSION_KEY,
+    });
     return await this.sendCommandToNode({
       url: urlWithPath,
       data: params.body,
